Successful

+++ Basic/Array/ft_printf.c
@@ -13,7 +13,7 @@ static int	ft_putstr(char *str)
 	return (len);
 }
 
-static int	ft_itoa_base_two(uintptr_t n, int base, char *str, int len)
+static int	ft_itoa_base_two(u_int64_t n, int base, char *str, int len)
 {
 	int	arr[100];
 	int	i;
@@ -33,7 +33,7 @@ static int	ft_itoa_base_two(uintptr_t n, int base, char *str, int len)
 	return (len);
 }
 
-static int	ft_itoa_base(uintptr_t n, int base, char *str, int mod)
+static int	ft_itoa_base(u_int64_t n, int base, char *str, int mod)
 {
 	int	len;
 
@@ -71,7 +71,7 @@ static int	ft_format(va_list args, char f)
 	else if (f == 'd' || f == 'i')
 		return (ft_itoa_base(va_arg(args, int), 10, DEC, 1));
 	else if (f == 'p')
-		return (ft_itoa_base(va_arg(args, uintptr_t), 16, HEXLOW, 2));
+		return (ft_itoa_base(va_arg(args, u_int64_t), 16, HEXLOW, 2));
 	else if (f == 'x')
 		return (ft_itoa_base(va_arg(args, unsigned int), 16, HEXLOW, 0));
 	else if (f == 'X')
